backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
requests==2.32.3
python-dotenv==1.1.1
google-generativeai==0.7.2
pdfminer.six==20240706
pydantic==2.8.2

